Avance login, pantallas clientes, proyectos y facturas

+++ recordatorio-pagos-back/src/main/java/com/recordatoriopagos/service/ClienteService.java
@@ -36,6 +36,9 @@ public class ClienteService {
 
 	public boolean eliminarUsuario(BigInteger idCliente) {
 		try {
+			if (clienteRepository.obtenerFacturasPorCliente(idCliente).size() > 0) {
+				return false;
+			}
 			clienteRepository.deleteById(idCliente);
 			return true;
 		} catch (Exception err) {
